# zstd: Inline nextFast call