```
benchmark                                               old ns/op      new ns/op      delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-32          4704626        4312341        -8.34%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-32      1068883        984396         -7.90%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-32       18021576       15813727       -12.25%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-32         13398648       11872693       -11.39%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-32       3892600        3691969        -5.15%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-32        5005336        4613696        -7.82%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-32           2099041        2012320        -4.13%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-32     189971         179829         -5.34%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-32     79873          80641          +0.96%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-32           14376218       12387421       -13.83%
BenchmarkDecoder_DecoderSmall/html.zst-32               1209031        1131450        -6.42%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-32      84458          79599          -5.75%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-32             586411         534015         -8.94%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-32         133502         123999         -7.12%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-32          1866430        1718598        -7.92%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-32            1407331        1282011        -8.90%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-32          484561         453559         -6.40%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-32           623368         573715         -7.97%
BenchmarkDecoder_DecodeAll/html_x_4.zst-32              260694         249948         -4.12%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-32        23118          21642          -6.38%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-32        9556           9489           -0.70%
BenchmarkDecoder_DecodeAll/urls.10K.zst-32              1627793        1500977        -7.79%
BenchmarkDecoder_DecodeAll/html.zst-32                  149442         139560         -6.61%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-32         10589          10084          -4.77%
BenchmarkDecoderSilesia-32                              403091267      354700567      -12.00%
BenchmarkDecoderEnwik9-32                               2119570000     1782801500     -15.89%

benchmark                                               old MB/s     new MB/s     speedup
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-32          313.43       341.94       1.09x
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-32      887.57       963.74       1.09x
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-32       213.90       243.77       1.14x
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-32         254.80       287.55       1.13x
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-32       257.27       271.25       1.05x
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-32        243.08       263.72       1.08x
BenchmarkDecoder_DecoderSmall/html_x_4.zst-32           1561.09      1628.37      1.04x
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-32     4312.23      4555.45      1.06x
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-32     12328.85     12211.44     0.99x
BenchmarkDecoder_DecoderSmall/urls.10K.zst-32           390.69       453.42       1.16x
BenchmarkDecoder_DecoderSmall/html.zst-32               677.57       724.03       1.07x
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-32      386.08       409.65       1.06x
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-32             314.32       345.16       1.10x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-32         888.28       956.37       1.08x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-32          258.17       280.38       1.09x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-32            303.24       332.88       1.10x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-32          258.33       275.99       1.07x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-32           243.98       265.09       1.09x
BenchmarkDecoder_DecodeAll/html_x_4.zst-32              1571.19      1638.74      1.04x
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-32        4429.42      4731.47      1.07x
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-32        12880.61     12972.79     1.01x
BenchmarkDecoder_DecodeAll/urls.10K.zst-32              431.31       467.75       1.08x
BenchmarkDecoder_DecodeAll/html.zst-32                  685.22       733.73       1.07x
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-32         384.93       404.20       1.05x
BenchmarkDecoderSilesia-32                              525.81       597.54       1.14x
BenchmarkDecoderEnwik9-32                               471.79       560.91       1.19x

```